Moved MbedTLS header inclusions into src file instead of pwcrypt.h
Added comfy build.sh

+++ build.sh
@@ -1,5 +1,5 @@
-#!/bin/sh
-#
+#!/bin/bash
+
 #  Copyright 2020 Raphael Beck
 #
 #  Licensed under the Apache License, Version 2.0 (the "License");
@@ -15,12 +15,35 @@
 #  limitations under the License.
 
 PROJECT_NAME="pwcrypt"
+
+if [ "$(whoami)" = "root" ]; then
+  echo "  Please don't run as root/using sudo..."
+  exit
+fi
+
+PREVCC="$CC"
+PREVCXX="$CXX"
+
+if command -v clang &> /dev/null
+then
+    echo "-- Clang found on system, great! Long live LLVM! :D"
+    export CC=clang
+    export CXX=clang++
+fi
+
 REPO=$(dirname "$0")
-cov=Off
-if [ "$1" = "cov" ]; then cov=On; fi
+rm -rf "$REPO"/out
 rm -rf "$REPO"/build
-mkdir -p "$REPO"/build && cd "$REPO"/build || exit
-cmake -DBUILD_SHARED_LIBS=Off -DUSE_SHARED_MBEDTLS_LIBRARY=Off "-D${PROJECT_NAME}_ENABLE_TESTS=On" "-D${PROJECT_NAME}_ONLY_BUILD_LIB=On" -DENABLE_COVERAGE="${cov}" -DCMAKE_BUILD_TYPE=Debug ..
-cmake --build . --config Debug || exit
-./run_tests || ./Debug/run_tests.exe || exit
+mkdir -p "$REPO"/build/include && cd "$REPO"/build || exit
+
+cmake -DBUILD_SHARED_LIBS=On -DUSE_SHARED_MBEDTLS_LIBRARY=Off "-D${PROJECT_NAME}_BUILD_DLL=On" "-D${PROJECT_NAME}_PACKAGE=On" -DCMAKE_BUILD_TYPE=Release ..
+
+cmake --build . --config Release
+
+export CC="$PREVCC"
+export CXX="$PREVCXX"
+
+cp -r ../include .
 cd "$REPO" || exit
+
+echo "  Done. Exported build into $REPO/build"
